+++ exchange_service/src/treasury_api_client/client.rs
@@ -7,22 +7,22 @@ pub struct TreasuryApiUrl(pub String);
 #[derive(Serialize, Deserialize, Debug)]
 pub struct RateRequestBody {
     //TODO: make it work with float
-    buy_amount: u32, //ethereum
+    buy_amount: u64, //ethereum
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct RateResponseBody {
     pub symbol: String,
-    pub rate: f32,
-    pub sell_amount: u32, //satoshis
-    pub buy_amount: u32,  //ethereum
+    pub rate: f64,
+    pub sell_amount: u64, //satoshis
+    pub buy_amount: u64,  //ethereum
 }
 
 pub trait ApiClient: Send + Sync {
     fn request_rate(
         &self,
         symbol: Symbol,
-        buy_amount: u32,
+        buy_amount: u64,
     ) -> Result<RateResponseBody, reqwest::Error>;
 }
 
@@ -36,7 +36,7 @@ impl ApiClient for DefaultApiClient {
     fn request_rate(
         &self,
         symbol: Symbol,
-        buy_amount: u32,
+        buy_amount: u64,
     ) -> Result<RateResponseBody, reqwest::Error> {
         let body = RateRequestBody { buy_amount };
